refactor: 마인드맵 constants 위치 변경
