+++ backend/api/src/main/java/com/yat2/episode/mindmap/constants/MindmapConstants.java
@@ -1,4 +1,4 @@
-package com.yat2.episode.global.constant;
+package com.yat2.episode.mindmap.constants;
 
 public class MindmapConstants {
     public static final String PRIVATE_NAME = "의 마인드맵";
